ddl, statistics: resolve the package dependency. (#4960)
