+++ executor/simple.go
@@ -21,7 +21,7 @@ import (
 	"github.com/juju/errors"
 	"github.com/pingcap/tidb/ast"
 	"github.com/pingcap/tidb/context"
-	"github.com/pingcap/tidb/ddl"
+	"github.com/pingcap/tidb/ddl/util"
 	"github.com/pingcap/tidb/infoschema"
 	"github.com/pingcap/tidb/model"
 	"github.com/pingcap/tidb/mysql"
@@ -322,6 +322,6 @@ func (e *SimpleExec) executeDropStats(s *ast.DropStatsStmt) error {
 		}
 		return errors.Trace(h.Update(GetInfoSchema(e.ctx)))
 	}
-	h.DDLEventCh() <- &ddl.Event{Tp: model.ActionDropTable, TableInfo: s.Table.TableInfo}
+	h.DDLEventCh() <- &util.Event{Tp: model.ActionDropTable, TableInfo: s.Table.TableInfo}
 	return nil
 }
